Replace old with new custom date
RISDEV-1444
